REST API v0, part 1 + some refactoring of models, repositories and services.

+++ src/main/java/com/aerotravel/flightticketbooking/repository/AirportRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AirportRepository extends JpaRepository<Airport, Integer> {
+public interface AirportRepository extends JpaRepository<Airport, Long> {
 }
